Apply default page template to group creation page
